Refactor NodePortTuple and add helper functions to OFMessageUtils
