+++ src/main/java/net/floodlightcontroller/topology/TopologyInstance.java
@@ -26,6 +26,7 @@ import java.util.Map;
 import java.util.PriorityQueue;
 import java.util.Set;
 
+import net.floodlightcontroller.core.types.NodePortTuple;
 import net.floodlightcontroller.routing.BroadcastTree;
 import net.floodlightcontroller.routing.Link;
 import net.floodlightcontroller.routing.Route;
